Redo ANTLR integration with local copy of headers
ANTLR integration has proven to be a pain, and I didn't plan to spend
the last three hours fixing linker errors. For now, I'm including a
local copy of the ANTLR runtime. I'll figure out how to properly link to
a system install or fetch the runtime in a package at some point in the
future.